+++ ChangeLog.rst
@@ -9,7 +9,10 @@ Old tests for widgets, removed from core framework were removed.
 
 Tests for widget's MOP methods were ported to Rove.
 
-Package ``weblocks/utils/close`` now loaded with main ``:weblocks`` system.
+Package ``weblocks/utils/close`` now loaded with main ``:weblocks``
+system.
+
+Few old widget tests were removed.
 
 0.25.0 (2018-01-31)
 ===================